refactor: production-ready error handling and security improvements
Major improvements:
- Replace all panic() with proper error handling using cockroachdb/errors
- Add unified error types with stack traces in lib/errs
- Add input validation (key/value size limits) in lib/validate
- Add constants package lib/consts for magic numbers
- Add buffer pool for performance optimization
- Update all server initialization to return errors instead of panic
- Upgrade version map from uint32 to uint64 to prevent overflow
- Update database interface to return errors instead of panic
- Update cluster module with proper error handling
- Update AOF rewrite with error returns
- Update logger to handle errors gracefully
- Maintain backward compatibility with Must* functions

New dependencies:
- github.com/cockroachdb/errors v1.12.0

Fixes: All panic() calls replaced with proper error handling

+++ database/transaction.go
@@ -27,7 +27,7 @@ func init() {
 }
 
 // invoker should lock watching keys
-func isWatchingChanged(db *DB, watching map[string]uint32) bool {
+func isWatchingChanged(db *DB, watching map[string]uint64) bool {
 	for key, ver := range watching {
 		currentVersion := db.GetVersion(key)
 		if ver != currentVersion {
